+++ include/OFIRM.h
@@ -296,7 +296,7 @@ public:
 
 	int      is_in_zoom_win();
 	int		find_settle_town();
-	int 		should_show_info();
+	int should_show_info() const;
 
 	int		set_builder(short newBuilderRecno);
 	int		find_idle_builder(int nearest);
@@ -308,7 +308,7 @@ public:
 	int		can_assign_capture();
 	int		can_worker_capture(int captureNationRecno);
 	virtual int	 is_worker_full();
-	int		have_own_workers(int checkSpy=0);
+	int have_own_workers(int checkSpy=0) const;
 
 	void 		set_worker_home_town(int townRecno, char remoteAction, int workerId=0);
 	int 		can_spy_bribe(int bribeWorkerId, int briberNationRecno);